Move some altair code to a separate file
